// src/types/Settings.ts
export default interface Settings {
	circleOutlineThickness: number;
	circleRadius: number;
	stroke: string;
	fill: string;
	pathColor: string;
	nodeHeight: number;
	nodeWidth: number;
	numNodes: number;
	labelFontSize: number;
	transitionDuration: number;
	titleType: "english" | "romaji" | "native";
}